pyproject.toml
[project]
name = "quantlab"
version = "0.1.0"
description = "Quantitative research and backtesting system"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Quant Researcher", email = "quant@example.com"}
]

keywords = ["quantitative", "trading", "backtest", "portfolio"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Data & Storage
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "polars>=0.20.0",
    "pyarrow>=15.0.0",
    "duckdb>=0.10.0",
    
    # Chinese Market Data
    "akshare>=1.15.0",
    
    # Configuration & Types
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "PyYAML>=6.0.1",
    
    # Time & Calendar
    "exchange-calendars>=4.5",
    "python-dateutil>=2.8.2",
    
    # Visualization
    "streamlit>=1.31.0",
    "plotly>=5.18.0",
    
    # Utilities
    "loguru>=0.7.2",
    "tqdm>=4.66.0",
    "joblib>=1.3.2",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
quantlab = "quantlab.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/quantlab"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
strict_optional = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/quantlab"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
]
